+++ classes/components/data/action/DoInterface.class.php
@@ -1,0 +1,7 @@
+<?php
+namespace Flex\Components\Data\Action;
+
+interface DoInterface{
+    public function do(?array $params=[]) : ?string;
+}
+?>